Refactor multiprocessing compatibility into general kolibri utils module.
