+++ kolibri/core/tasks/worker.py
@@ -3,11 +3,11 @@ from concurrent.futures import CancelledError
 
 from django.db import connection as django_connection
 
-from kolibri.core.tasks.compat import PoolExecutor
 from kolibri.core.tasks.constants import Priority
 from kolibri.core.tasks.storage import Storage
 from kolibri.core.tasks.utils import db_connection
 from kolibri.core.tasks.utils import InfiniteLoopThread
+from kolibri.utils.multiprocessing_compat import PoolExecutor
 
 logger = logging.getLogger(__name__)
